cleanup

+++ in-image-script.sh
@@ -847,33 +847,35 @@ startfunc
 
    # Don't remake debs if they already exist in output.
    KERNEL_VERS=$(< /tmp/KERNEL_VERS)
-   [[ ! $REBUILD ]] && if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-image-*"${KERNEL_VERS}"* -print -quit)"
-   then
-        echo -e "${KERNEL_VERS} linux image on cache volume. 😎\n"
-        cp "${apt_cache}"/linux-image-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
-        echo "linux-image" >> /tmp/nodebs
-    elif if test -n "$(find /output/ -maxdepth 1 -name linux-image-*"${KERNEL_VERS}"* -print -quit)"
-   then
-        echo -e "${KERNEL_VERS} linux image found in /output/. 😎\n"
-        cp /output/linux-image-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
-        echo "linux-image" >> /tmp/nodebs
-    else
-        rm -f /tmp/nodebs || true
+   if [[ ! $REBUILD ]]
+        then
+            if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-image-*"${KERNEL_VERS}"* -print -quit)"
+            then
+                echo -e "${KERNEL_VERS} linux image on cache volume. 😎\n"
+                cp "${apt_cache}"/linux-image-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
+                echo "linux-image" >> /tmp/nodebs
+            elif if test -n "$(find /output/ -maxdepth 1 -name linux-image-*"${KERNEL_VERS}"* -print -quit)"
+            then
+                echo -e "${KERNEL_VERS} linux image found in /output/. 😎\n"
+                cp /output/linux-image-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
+                echo "linux-image" >> /tmp/nodebs
+                else
+                rm -f /tmp/nodebs || true
+            fi
+            if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-headers-*"${KERNEL_VERS}"* -print -quit)"
+            then
+                echo -e "${KERNEL_VERS} linux headers on cache volume. 😎\n"
+                cp "${apt_cache}"/linux-headers-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
+                echo "linux-image" >> /tmp/nodebs
+            elif if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-headers-*"${KERNEL_VERS}"* -print -quit)"
+            then
+                echo -e "${KERNEL_VERS} linux headers found in /output/. 😎\n"
+                cp /output/linux-headers-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
+                echo "linux-image" >> /tmp/nodebs
+                else
+                rm -f /tmp/nodebs || true
+            fi
     fi
-    [[ ! $REBUILD ]] && if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-headers-*"${KERNEL_VERS}"* -print -quit)"
-   then
-        echo -e "${KERNEL_VERS} linux headers on cache volume. 😎\n"
-        cp "${apt_cache}"/linux-headers-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
-        echo "linux-image" >> /tmp/nodebs
-    elif if test -n "$(find "${apt_cache}" -maxdepth 1 -name linux-headers-*"${KERNEL_VERS}"* -print -quit)"
-   then
-        echo -e "${KERNEL_VERS} linux headers found in /output/. 😎\n"
-        cp /output/linux-headers-*"${KERNEL_VERS}"*arm64.deb "${workdir}"/
-        echo "linux-image" >> /tmp/nodebs
-    else
-        rm -f /tmp/nodebs || true
-    fi
-    
     [[ $REBUILD ]] && rm -f /tmp/nodebs || true
     
     if [[ -e /tmp/nodebs ]]
